Fixes a race condition found in percept_db start/1 function.
When function start/1 finds an instance of percept_db running, it will send a message to stop it and spawn a process to start a new one, which leads to a race condition.
The function will return {restarted, pid()} but the pid() will die once it tries to create the ets table.